divide Calculator into separate components for better clarity
