+++ src/components/Calculator/UserInput.jsx
@@ -1,0 +1,14 @@
+function UserInput({calculateFonts}) {
+  return (
+    <div className="w-full h-gr2 bg-gold">
+      <input
+        type="number"
+        onChange={(e) => calculateFonts(parseInt(e.target.value))}
+        placeholder="Enter your biggest font size"
+        className="p-2"
+      />
+    </div>
+  )
+}
+
+export default UserInput;
